platforms/astbmc: Create separate inits for the AST2400 and AST2500
While the AST2400 and AST2500 are very close in their designs, some
aspects are different. This can easily be handled by platforms using a
different init function which can deal with the differences, most of the
code will be common.

This patch performs the initial split and will init the SPI controller
for the PNOR correctly for the two chips.