libflash/blocklevel: Optionally always return opal-api return codes
With the introduction of the blocklevel abstraction layer some of the
return code passing got neglected. This isn't a very big deal for
skiboot/external tools however this can be a problem in skiboot proper.
Some of the opal-api calls could pass back values from blocklevel
directly along the api in which case it is quite important that the
values are valid and make some sense.