libflash/mbox-flash: Implement MARK_WRITE_ERASED mbox call
Version two of the mbox-flash protocol defines a new command:
MARK_WRITE_ERASED.

This command provides a simple way to mark a region of flash as all 0xff
without the need to go and write all 0xff. This is an optimisation as
there is no need for an erase before a write, it is the responsibility of
the BMC to deal with the flash correctly, however in v1 it was ambiguous
what a client should do if the flash should be erased but not actually
written to. This allows of a optimal path to resolve this problem.