libflash/mbox-flash: Handle BMC reboots
If a BMC reboots it is expected that it will raise a bit in the ATTN
register. On seeing this bit skiboot should assume that the mbox daemon
on the BMC is gone and that it should assume nothing exists anymore.
Critically it must also resent a GET_MBOX_INFO info so that the
respawned daemon knows what version of the protocol to talk and so that
there can be no confusion as to 'block size'.