external/gard: Stop coping with ECC
Recent work with the blocklevel abstraction means that consumers of
blocklevel can safely call blocklevel_read() and blocklevel_write() and
as long as the blocklevel knows of the presence of ECC then it will deal
with all cases.

This commit removes code in the gard tool which compensated for
inadequacies no longer present in blocklevel.