libflash/mbox-flash: Handle windows going bad
For various reasons the BMC may inform us that the data in our currently
open windows may be bad. If this happens we are expected to wait until
it confirms. For now skiboot will simply fail read/writes/erases if it
detects any of these condition, expected behaviour is that the caller
will retry. If this happens during boot skiboot and this causes a read
of (for example) BOOTKERNEL to fail then the system will likely fail to
boot. The BMC should be aware that the system is booting and that making
the windows unreliable is unrecoverable during this time.