external/pflash: Fix erasing within a single erase block
It is possible to erase within a single erase block. Currently the
pflash code assumes that if the erase starts part way into an erase
block it is because it needs to be aligned up to the boundary with the
next erase block.

Doing an erase smaller than a single erase block will cause underflows
and looping forever on erase.

Fixes: ae6cb86c2 ("external/pflash: Reinstate the progress bars")

---
Side note, this bug is only possible on 64K erase granule platforms.
Also, this bug cannot be caught by our test suite as it stands because
the test suite tests with the file backend to blocklevel and it has an
erase granule of one. I'll have a think about perhaps adding changes so
that we might be able to catch this.